requirements.txt
aiogram
motor
aiohttp
yt-dlp
python-dotenv